Refactoring as_dict method:
Refactor as_dict method to new class and import it
experiments.py and dataItems.py files
