+++ experiment_server/models/DictionaryCreator.py
@@ -1,0 +1,9 @@
+class DictionaryCreator:
+    def as_dict(self):
+        result = {}
+        for c in self.__table__.columns:
+            if c.name == 'startDatetime' or c.name == 'endDatetime':
+                result[c.name] = str(getattr(self, c.name))
+            else:
+                result[c.name] = getattr(self, c.name)
+        return result
