+++ experiment_server/models/dataitems.py
@@ -7,9 +7,10 @@ from sqlalchemy import (
     PickleType
 )
 from .meta import Base
+from .DictionaryCreator import DictionaryCreator
 
 
-class DataItem(Base):
+class DataItem(Base, DictionaryCreator):
     __tablename__ = 'dataitems'
     id = Column(Integer, primary_key=True)
     user_id = Column(Integer, ForeignKey('users.id'))
@@ -17,12 +18,3 @@ class DataItem(Base):
     value = Column(PickleType)
     startDatetime = Column(DateTime)
     endDatetime = Column(DateTime)
-
-    def as_dict(self):
-        result = {}
-        for c in self.__table__.columns:
-            if c.name == 'startDatetime' or c.name == 'endDatetime':
-                result[c.name] = str(getattr(self, c.name))
-            else:
-                result[c.name] = getattr(self, c.name)
-        return result
